% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligera2.R
\name{ligera2}
\alias{ligera2}
\title{LIGERA2: LIght GEnetic Robust Association main function}
\usage{
ligera2(
  X,
  trait,
  mean_kinship,
  loci_on_cols = FALSE,
  mem_factor = 0.7,
  mem_lim = NA,
  tol = 1e-15,
  verbose = FALSE
)
}
\arguments{
\item{X}{The \code{m}-by-\code{n} genotype matrix, containing dosage values in (0, 1, 2, NA) for the reference allele at each locus.}

\item{trait}{The length-\code{n} trait vector, which may be real valued and contain missing values.}

\item{mean_kinship}{An estimate of the mean kinship produced externally, to ensure internal estimates of kinship and inbreeding are unbiased.}

\item{loci_on_cols}{If \code{TRUE}, \code{X} has loci on columns and individuals on rows; if false (the default), loci are on rows and individuals on columns.
If \code{X} is a BEDMatrix object, \code{loci_on_cols = TRUE} is set automatically.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{tol}{Tolerance value passed to conjugate gradient method solver.}

\item{verbose}{If TRUE, reports on iterations progress (for debugging)}
}
\value{
A tibble containing the following association statistics
\itemize{
\item \code{pval}: The p-value of the association test
\item \code{beta}: The estimated effect size coefficient for the trait vector at this locus
\item \code{beta_std_dev}: The estimated coefficient variance of this locus (varies due to dependence on minor allele frequency)
\item \code{p_q}: The allele variance estimate (estimate of \code{p*(1-p)}).  The number of heterozygotes, weighted by inbreeding coefficient, and with pseudocounts included, is used in this estimate (in other words, it does not equal MAF * ( 1 - MAF ), where MAF is the marginal allele frequency.
\item \code{t_stat}: The test statistic, equal to \code{beta / beta_std_dev}.
}
}
\description{
This function performs the genetic association tests on every locus of a genotype matrix against a quantitative trait, implicitly computing the kinship matrix in a way that scales better than an explicit kinship estimate.
The function returns a tibble containing association statistics and several intermediates.
}
\details{
Suppose there are \code{n} individuals and \code{m} loci.
}
\examples{
# Construct toy data
# genotype matrix
X <- matrix(
    c(0, 1, 2,
      1, 0, 1,
      1, 0, 2),
    nrow = 3,
    byrow = TRUE
)
trait <- 1 : 3
mean_kinship <- mean( diag( 3 ) / 2 ) # unstructured case

tib <- ligera2( X, trait, mean_kinship )
tib

}
\seealso{
The \code{popkin} package.
}
