% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligera2_bed_f.R
\name{ligera2_bed_f}
\alias{ligera2_bed_f}
\title{LIGERA2 BED: LIght GEnetic Robust Association main function}
\usage{
ligera2_bed_f(
  file,
  m_loci,
  n_ind,
  trait,
  mean_kinship,
  covar = NULL,
  mem_factor = 0.7,
  mem_lim = NA,
  m_chunk_max = 1000,
  V = 0,
  tol = 1e-15
)
}
\arguments{
\item{file}{The path to the BED file containing the genotypes, potentially excluding the BED extension.}

\item{m_loci}{The number of loci in the BED file.}

\item{n_ind}{The number of individuals in the BED file.}

\item{trait}{The length-\code{n} trait vector, which may be real valued and contain missing values.}

\item{mean_kinship}{An estimate of the mean kinship produced externally, to ensure internal estimates of kinship are unbiased.}

\item{covar}{An optional \code{n}-by-\code{K} matrix of \code{K} covariates, aligned with the individuals.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{m_chunk_max}{Sets the maximum number of loci to process at the time.
Actual number of loci loaded may be lower if memory is limiting.}

\item{V}{Algorithm version (0, 1, 2).
Experimental features, not worth explaining.}

\item{tol}{Tolerance value passed to conjugate gradient method solver.}
}
\value{
A tibble containing the following association statistics
\itemize{
\item \code{pval}: The p-value of the association test
\item \code{beta}: The estimated effect size coefficient for the trait vector at this locus
\item \code{f_stat}: The F statistic
\item \code{df}: degrees of freedom: number of non-missing individuals minus number of parameters of full model
}
}
\description{
This function performs the genetic association tests on every locus of a genotype matrix against a quantitative trait, implicitly computing the kinship matrix in a way that scales better than an explicit kinship estimate.
The function returns a tibble containing association statistics and several intermediates.
This version calculates p-values using an F-test, which gives calibrated statistics under both quantitative and binary traits.
Compared to \code{\link[=ligera2]{ligera2()}}, which uses the faster Wald test (calibrated for quantitative but not binary traits), this F-test version is quite a bit slower, and is optimized for \verb{m >> n}, so it is a work in progress.
This optimized version requires the genotypes to be in a file in BED format.
}
\details{
Suppose there are \code{n} individuals and \code{m} loci.
}
\examples{
# MISSING SAMPLE BED FILE

}
\seealso{
The \code{popkin} package.
}
