% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligera2_bed_multi.R
\name{ligera2_bed_multi}
\alias{ligera2_bed_multi}
\title{LIGERA2 BED multi: LIght GEnetic Robust Association multiscan function}
\usage{
ligera2_bed_multi(
  file,
  m_loci,
  n_ind,
  trait,
  mean_kinship,
  q_cut = 0.05,
  one_per_iter = FALSE,
  covar = NULL,
  mem_factor = 0.7,
  mem_lim = NA,
  m_chunk_max = 1000,
  tol = 1e-15
)
}
\arguments{
\item{file}{The path to the BED file containing the genotypes, potentially excluding the BED extension.}

\item{m_loci}{The number of loci in the BED file.}

\item{n_ind}{The number of individuals in the BED file.}

\item{trait}{The length-\code{n} trait vector, which may be real valued and contain missing values.}

\item{mean_kinship}{An estimate of the mean kinship produced externally, to ensure internal estimates of kinship and inbreeding are unbiased.}

\item{q_cut}{The q-value threshold to admit new loci into the polygenic model.}

\item{one_per_iter}{If true, only the most significant locus per iteration is added to model of next iteration.  Otherwise all significant loci per iteration are added to the model of next iteration.}

\item{covar}{An optional \code{n}-by-\code{K} matrix of \code{K} covariates, aligned with the individuals.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{m_chunk_max}{Sets the maximum number of loci to process at the time.
Actual number of loci loaded may be lower if memory is limiting.}

\item{tol}{Tolerance value passed to conjugate gradient method solver.}
}
\value{
A tibble containing the following association statistics from the last scan for non-selected loci.
For selected loci, these are the values from the scan before each was added to the model (as after addition they get \verb{beta ~= 0} and \verb{pval ~= 1}).
\itemize{
\item \code{pval}: The p-value of the last association scan.
\item \code{beta}: The estimated effect size coefficient for the trait vector at this locus.
\item \code{beta_std_dev}: The estimated coefficient variance of this locus (varies due to dependence on minor allele frequency).
\item \code{p_q}: The allele variance estimate (estimate of \code{p*(1-p)}).  The number of heterozygotes, weighted by inbreeding coefficient, and with pseudocounts included, is used in this estimate (in other words, it does not equal MAF * ( 1 - MAF ), where MAF is the marginal allele frequency.
\item \code{t_stat}: The test statistic, equal to \code{beta / beta_std_dev}.
\item \code{qval}: The q-value of the last association scan.
\item \code{sel}: the order in which loci were selected, or zero if they were not selected.
}
}
\description{
This function performs multiple genetic association scans, adding one significant locus per iteration to the model (modeled as a covariate) to increase power in the final model.
The function returns a tibble containing association statistics and several intermediates.
This optimized version requires the genotypes to be in a file in BED format.
}
\details{
Suppose there are \code{n} individuals and \code{m} loci.
}
\examples{
# MISSING SAMPLE BED FILE

}
\seealso{
The \code{popkin} package.
}
