% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligera2_bed_trans.R
\name{ligera2_bed_trans}
\alias{ligera2_bed_trans}
\title{LIGERA2 BED: LIght GEnetic Robust Association main function}
\usage{
ligera2_bed_trans(
  file,
  m_loci,
  n_ind,
  trait,
  mean_kinship,
  covar = NULL,
  mem_factor = 0.7,
  mem_lim = 1,
  tol = 1e-15
)
}
\arguments{
\item{file}{The path to the BED file containing the genotypes, potentially excluding the BED extension.}

\item{m_loci}{The number of loci in the BED file.}

\item{n_ind}{The number of individuals in the BED file.}

\item{trait}{The length-\code{n} trait vector, which may be real valued and contain missing values.}

\item{mean_kinship}{An estimate of the mean kinship produced externally, to ensure internal estimates of kinship and inbreeding are unbiased.}

\item{covar}{An optional \code{n}-by-\code{K} matrix of \code{K} covariates, aligned with the individuals.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{tol}{Tolerance value passed to conjugate gradient method solver.}
}
\value{
A tibble containing the following association statistics
\itemize{
\item \code{pval}: The p-value of the association test
\item \code{beta}: The estimated effect size coefficient for the trait vector at this locus
\item \code{beta_std_dev}: The estimated coefficient variance of this locus (varies due to dependence on minor allele frequency)
\item \code{p_q}: The allele variance estimate (estimate of \code{p*(1-p)}).  The number of heterozygotes, weighted by inbreeding coefficient, and with pseudocounts included, is used in this estimate (in other words, it does not equal MAF * ( 1 - MAF ), where MAF is the marginal allele frequency.
\item \code{t_stat}: The test statistic, equal to \code{beta / beta_std_dev}.
}
}
\description{
DOUBLE TRANSPOSE METHOD - This function performs the genetic association tests on every locus of a genotype matrix against a quantitative trait, implicitly computing the kinship matrix in a way that scales better than an explicit kinship estimate.
The function returns a tibble containing association statistics and several intermediates.
This optimized version requires the genotypes to be in a file in BED format.
}
\details{
Suppose there are \code{n} individuals and \code{m} loci.
}
\examples{
# MISSING SAMPLE BED FILE

}
\seealso{
The \code{popkin} package.
}
