% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligera2_f.R
\name{ligera2_f}
\alias{ligera2_f}
\title{LIGERA2: LIght GEnetic Robust Association main function}
\usage{
ligera2_f(
  X,
  trait,
  mean_kinship,
  covar = NULL,
  loci_on_cols = FALSE,
  mem_factor = 0.7,
  mem_lim = NA,
  m_chunk_max = 1000,
  V = 0,
  tol = 1e-15
)
}
\arguments{
\item{X}{The \code{m}-by-\code{n} genotype matrix, containing dosage values in (0, 1, 2, NA) for the reference allele at each locus.}

\item{trait}{The length-\code{n} trait vector, which may be real valued and contain missing values.}

\item{mean_kinship}{An estimate of the mean kinship produced externally, to ensure internal estimates of kinship are unbiased.}

\item{covar}{An optional \code{n}-by-\code{K} matrix of \code{K} covariates, aligned with the individuals.}

\item{loci_on_cols}{If \code{TRUE}, \code{X} has loci on columns and individuals on rows; if false (the default), loci are on rows and individuals on columns.
If \code{X} is a BEDMatrix object, \code{loci_on_cols = TRUE} is set automatically.}

\item{mem_factor}{Proportion of available memory to use loading and processing genotypes.
Ignored if \code{mem_lim} is not \code{NA}.}

\item{mem_lim}{Memory limit in GB, used to break up genotype data into chunks for very large datasets.
Note memory usage is somewhat underestimated and is not controlled strictly.
Default in Linux and Windows is \code{mem_factor} times the free system memory, otherwise it is 1GB (OSX and other systems).}

\item{m_chunk_max}{Sets the maximum number of loci to process at the time.
Actual number of loci loaded may be lower if memory is limiting.}

\item{V}{Algorithm version (0, 1, 2, 3).
Experimental features, not worth explaining except that V=1 is numerically least stable but also the fastest (for n << m).}

\item{tol}{Tolerance value passed to conjugate gradient method solver.}
}
\value{
A tibble containing the following association statistics
\itemize{
\item \code{pval}: The p-value of the association test
\item \code{beta}: The estimated effect size coefficient for the trait vector at this locus
\item \code{f_stat}: The F statistic
\item \code{df}: degrees of freedom: number of non-missing individuals minus number of parameters of full model
}
}
\description{
This function performs the genetic association tests on every locus of a genotype matrix against a quantitative trait, implicitly computing the kinship matrix in a way that scales better than an explicit kinship estimate.
The function returns a tibble containing association statistics and several intermediates.
This version calculates p-values using an F-test, which gives calibrated statistics under both quantitative and binary traits.
Compared to \code{\link[=ligera2]{ligera2()}}, which uses the faster Wald test (calibrated for quantitative but not binary traits), this F-test version is quite a bit slower, and is optimized for \verb{m >> n}, so it is a work in progress.
}
\details{
Suppose there are \code{n} individuals and \code{m} loci.
}
\examples{
# Construct toy data
# genotype matrix
X <- matrix(
    c(0, 1, 2,
      1, 0, 1,
      1, 0, 2),
    nrow = 3,
    byrow = TRUE
)
trait <- 1 : 3
mean_kinship <- mean( diag( 3 ) / 2 ) # unstructured case

tib <- ligera2_f( X, trait, mean_kinship )
tib

}
\seealso{
The \code{popkin} package.
}
